Fix handling of default/empty log level
Set $verbose to an empty string if no acceptable log level was passed. Then
unquote the resulting string in the exec command so that it can expand to
nothing. The contents are already sanitised so we shouldn't need to worry
about quoting any value.

Added examples of the three "known" logging levels that are parsed.

With the default configuration of `LOGSTASH_LOGLEVEL=warn` the wrapper
script was not setting the $verbose variable. The resulting command then
featured a pair of quotes with no contents inside: `""`

The logstash JAR rejected this with:
```
No such command ""
Available commands:
  -v
  -V
  --version
  agent
  web
  test
  irb
  pry
```